requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
